WIP: Refactor integration of scripting from AvaloniaHostApp to HostApp base class.
