+++ src/apps/Avalonia/Highbyte.DotNet6502.App.Avalonia.Desktop/Program.cs
@@ -19,6 +19,7 @@ using System.Diagnostics;
 using Highbyte.DotNet6502.DebugAdapter;
 using Highbyte.DotNet6502.Scripting;
 using Highbyte.DotNet6502.Scripting.MoonSharp;
+using Highbyte.DotNet6502.Systems;
 
 namespace Highbyte.DotNet6502.App.Avalonia.Desktop;
