+++ src/libraries/Highbyte.DotNet6502.Scripting.MoonSharp/MoonSharpScriptingConfigurator.cs
@@ -1,3 +1,4 @@
+using Highbyte.DotNet6502.Systems;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Logging;
